Move test running embeded etcd to separate package
